[backend] combined mongo and server so they started together
