+++ backend/src/app.ts
@@ -1,21 +1,45 @@
-/**
- * Initializes mongoose and express.
- */
+import dotenv from "dotenv";
+import express, { Express, NextFunction, Request, Response } from "express";
+import { isHttpError } from "http-errors";
 
-import mongoose from "mongoose";
+import announcementRoutes from "./routes/announcement";
+import discussionRoutes from "./routes/discussion";
+import replyRoutes from "./routes/reply";
+import userRoutes from "./routes/user";
 
-import { mongoURI } from "./config";
+dotenv.config();
 
-// Connect to MongoDB
-void mongoose
-  .connect(mongoURI)
-  .then(() => {
-    console.log("Connected to Database.");
-  })
-  .catch((error: unknown) => {
-    if (error instanceof Error) {
-      console.log(error.message);
-    } else {
-      console.log("An unknown error occurred");
-    }
-  });
+const app: Express = express();
+
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+app.get("/", (req: Request, res: Response) => {
+  res.send("Express + TypeScript Server");
+});
+
+app.use("/api/users", userRoutes);
+app.use("/api/announcement", announcementRoutes);
+app.use("/api/discussions", discussionRoutes);
+app.use("/api/replies", replyRoutes);
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+app.use((error: unknown, req: Request, res: Response, next: NextFunction) => {
+  // 500 is the "internal server error" error code, this will be our fallback
+  let statusCode = 500;
+  let errorMessage = "An error has occurred.";
+
+  // check is necessary because anything can be thrown, type is not guaranteed
+  if (isHttpError(error)) {
+    // error.status is unique to the http error class, it allows us to pass status codes with errors
+    statusCode = error.status;
+    errorMessage = error.message;
+  }
+  // prefer custom http errors but if they don't exist, fallback to default
+  else if (error instanceof Error) {
+    errorMessage = error.message;
+  }
+
+  res.status(statusCode).json({ error: errorMessage });
+});
+
+export default app;
