+++ backend/src/routes/user.ts
@@ -18,11 +18,11 @@ router.delete("/:id", UserValidator.deleteUser, UserController.deleteUser);
 router.get("/", UserController.getAllUsers);
 router.get("/:id", UserValidator.getUser, UserController.getUser);
 
-router.get("/personal-information", UserController.getPersonalInformation);
-router.post("/personal-information", UserController.editPersonalInformation); 
+router.get("/profile/personal-information", UserController.getPersonalInformation);
+router.post("/profile/personal-information", UserController.editPersonalInformation);
 
-router.get("/professional-information", UserController.getProfessionalInformation);
-router.post("/professional-information", UserController.editProfessionalInformation);
+router.get("/profile/professional-information", UserController.getProfessionalInformation);
+router.post("/profile/professional-information", UserController.editProfessionalInformation);
 
 router.get("/directory/personal-information", UserController.getDirectoryPersonalInformation);
 router.post("/directory/personal-information", UserController.editDirectoryPersonalInformation);
